fix(playerbot): log the attack skill's target vid, not an address
GetVID() returns a const VID&, and handed to a %u through sys_log's
varargs as it is, the class is never converted to the number it stands
for. Every "used attack skill" line printed the same target_vid -
4294758416, 0xFFFCD010, which reads as a stack address - for five bots
on three maps through a whole morning of traces, while the fight lines
around it named different monsters. Cast it, as every other line in the
overlay that prints a vid already does.

Log only; no behaviour changes. Syntax-checked against port40250.

+++ linux-port/overlays/playerbot/src/game/src/playerbot_combat.h
@@ -402,7 +402,7 @@ namespace
 						 : PLAYERBOT_SKILL_ATTACK_INTERVAL);
 				state.dwNextAttackTime = dwNow + PLAYERBOT_SKILL_ANIMATION_LOCK;
 				sys_log(0, "PLAYERBOT_AI: used attack skill pid=%u name=%s vnum=%u target_vid=%u",
-						ch->GetPlayerID(), ch->GetName(), skillVnum, target->GetVID());
+						ch->GetPlayerID(), ch->GetName(), skillVnum, (DWORD)target->GetVID());
 				return true;
 			}
 		}
